refactor PaymentField ts
